// FE/Web/Bloodline-DNA/src/features/manager/types/sample-ins.ts
// types/sample-ins.ts

export interface SampleInsRequest {
  sampleType: number;
  instructionText: string;
  mediaUrl: string;
}

export interface SampleInsResponse {
  id: number;
  sampleType: number;
  instructionText: string;
  mediaUrl: string;
}

export interface SampleInsUpdateRequest extends SampleInsRequest {
  id: number;
}